Isolate memory_cleaning tests
