+++ crates/burn-backend-tests/tests/fusion/mod.rs
@@ -4,7 +4,6 @@ pub use super::*;
 // mod fusion_f16_broadcast;
 // mod fusion_f16_write_vectorization;
 mod fusion_shape;
-// mod memory_cleaning;
 mod reduce_broadcasted;
 
 use burn_tensor::StreamId;
